requirements.txt
Django==5.0.1
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.27.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
